Telegram producer/FunctionApp/requirements.txt
azure-functions==1.18.0
pymongo==4.12.0
Telethon==1.38.1
